==> 5A_HTT_Leavitt_PCA.Rmd <==
HTT_Leavitt Principal Component Analysis
========================================================
## By: Sumaiya Islam
## Date: December 15, 2014

### Script contents:
- PCA on the cortex-only dataset (before and after CETS brain-cell type correction)
- PCA on the matched dataset (before and after CETS brain-cell type correction)

### A. Set up working directory & packages

Use R.2.14

We will initially set our working directory and load our libraries.
```{r HTT_Leavitt_load_libraries, include=FALSE, echo=FALSE}
setwd("/home/sislam/HTT_Leavitt")
library(methylumi)
library(gplots)
library(marray)
library(lumi)
library(lattice)
library("RColorBrewer")
library(knitr)
library(xtable)
# source("http://bioconductor.org/biocLite.R")
# biocLite("qvalue")
library(qvalue)
library(cets)
library(reshape)
library(ggplot2)
library(RColorBrewer)
library(grid)
library(gridExtra)
```

### B. Load files

#### We will be analyzing each of the subsetted datasets generated from the normalization script


```{r HTT_Leavitt_fileLoad, echo=FALSE}

# cortex-only dataset
## cortex-only dataset (normalized, ComBat-corrected, BEFORE CETS cell type correction)
load("HTT_Leavitt.cortex.combat.RData")
cortexDat.uncor<-HTT_Leavitt.cortex.combat
meta<-read.csv(file="cortexDes.final.csv", header=T, row.names=1)
identical(rownames(meta), rownames(pData(cortexDat.uncor)))
cortexDes.uncor<-meta
identical(colnames(exprs(cortexDat.uncor)),rownames(cortexDes.uncor))

## cortex-only dataset (normalized, ComBat-corrected, AFTER CETS cell type correction)
load("HTT_Leavitt.cortex.cor.final.RData") 
cortexDat.cor<-HTT_Leavitt.cortex.cor
identical(rownames(meta), rownames(pData(cortexDat.cor)))
cortexDes.cor<-meta
identical(colnames(exprs(cortexDat.cor)),rownames(cortexDes.cor))

# matched samples
## matched dataset (normalized, BEFORE CETS cell type correction)
load("HTT_Leavitt.Matched.norm.RData")
matchedDat.uncor<-HTT_Leavitt.Matched.swan
matched.meta<-read.csv(file="matchedDes.final.csv", header=T, row.names=1)
matched.meta<-matched.meta[match(colnames(matchedDat.uncor), rownames(matched.meta)),]
identical(rownames(matched.meta), rownames(pData(matchedDat.uncor)))
matchedDes.uncor<-matched.meta
identical(colnames(exprs(matchedDat.uncor)),rownames(matchedDes.uncor))

## matched dataset (normalized, AFTER CETS cell type correction)
load("HTT_Leavitt.Matched.cor.final.RData")
matchedDat.cor<-HTT_Leavitt.Matched.cor
matched.meta<-matched.meta[match(colnames(matchedDat.cor), rownames(matched.meta)),]
matchedDes.cor<-matched.meta
identical(colnames(exprs(matchedDat.cor)),rownames(matchedDes.cor))
```

### C. Principal Component Analysis (PCA)

#### PCA is a data dimensionality reduction method which attempts to define linearly independent principal components which explain incremental proportions of variance of the dataset. Measured variables in the metadata can be correlated to these principal components to discern which degree of variation they describe. A more detailed description of PCA and its assumptions are provided below: 

Principal components analysis is a popular tool for studying high-dimensional data. It relies on four major assumptions:
1. Linearity. This means that the only interaction among different signal sources is that they add together. If the strength of a combined signal were the product of the strengths of contributing signals, for instance, this would be a non-linear interaction and PCA would not work.
2. The interesting dynamics have the largest variances.
3. Mean and variance are sufficient statistics. Since PCA is designed around the covariance matrix of mean-centered data, the only statistics it considers are the mean and variance. If the data cannot adequately be described by its mean and variance (e.g. it is not Gaussian or exponentially distributed), PCA will be inadequate.
4. Orthogonal components. This is a particularly strong assumption made by PCA. It is this assumption that allows PCA to be computationally straightforward, but is not a realistic assumption in many data sets.

#### We will perform principal component analysis (PCA) to see if any of the covariates in the metadata correlate with the principal components. PCA help us investigate the major sources of variation in the dataset. Based on the preliminary plots generated during preprocessing, we already can tell that PC1 likely comprises methylation differences due to tissue type. What I'd like to do is to correlate the PC1 score to each probe and see which probes contribute the most to PC1, thereby identifying the CpG sites that explain most of the methylation variation due to tissue type.

### PCA on Cortex-Only Dataset

#### For normalized, ComBat-corrected, BEFORE CETS-correction cortex-only dataset:

```{r echo=FALSE}
pcs.cortex<-prcomp(exprs(cortexDat.uncor), center=TRUE, scale=FALSE) # scaling is not necessary for normalized datasets
```

```{r echo=FALSE}
plot(pcs.cortex)
```

The figure shown is a scree plot which depicts the proportion of variance in the dataset that are explained by each of the principal components. We can see that the first principal component explains most of the variance in the methylation dataset. 

```{r echo=FALSE}
(pc.cortex.sum<-summary(pcs.cortex))
```

We can see that PC1 explains 99.41% of the variance of the dataset.  However, we are only interested in the remaining PCs since the first PC just represents probe offset. We can re-adjust the proportion of variance explained by all subsequent PCs as follows: 


```{r echo=FALSE}
sum.prop<-pc.cortex.sum$importance
top.PC.prop<-sum.prop[2,1]
adj.sum.prop<-apply(sum.prop, 1, function(x) x/(1-top.PC.prop))
adj.prop.var<-adj.sum.prop[2:ncol(sum.prop),2]
adj.prop.var<-data.frame(adj.prop.var)
adj.prop.var$PC<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
rownames(adj.prop.var)<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
adj.prop.var$PC <- factor(adj.prop.var$PC, levels = paste("PC", 1:(ncol(sum.prop)-1), sep=""))
ggplot(adj.prop.var, aes(x=PC, y=adj.prop.var)) + geom_bar(stat="identity",fill="purple") + theme_bw() + ylab("Proportion of Variance")
```

We will now correlate the known covariates with the PCs (note: we will also include the neuronal cell proportions calculated in the CETS brain cell type correction script). 

```{r echo=FALSE}
# convert categorical data into factors => do ANOVA
cortexDes.uncor$Sentrix_ID <- as.factor(cortexDes.uncor$Sentrix_ID)
cortexDes.uncor$Sentrix_Position <- as.character(cortexDes.uncor$Sentrix_Position)
cortexDes.uncor$Sample_Group <- as.factor(cortexDes.uncor$Sample_Group)
cortexDes.uncor$Sex<- as.factor(cortexDes.uncor$Sex)
cortexDes.uncor$Disease_status <- as.factor(cortexDes.uncor$Disease_status)
prop<-read.table("Cortex_CETS_Neuronal_Glia_Proportions.txt") # include neuron/glia proportions into meta data
identical(rownames(cortexDes.uncor),rownames(prop))
cortexDes.uncor$neuron<-prop$neuron
cortexDes.uncor$glia<-prop$glia
# for Row
for (i in 1:nrow(cortexDes.uncor)){
  cortexDes.uncor$Row[i]<-paste(substr(cortexDes.uncor[i,"Sentrix_Position"], start=1, stop=3))
}
cortexDes.uncor$Row<- as.factor(cortexDes.uncor$Row)
str(cortexDes.uncor)

# continuous data => correlation (do spearman correlation because data is not always normally distributed)
```

Here are the eigenvectors (principal components) for the first few samples in the dataset:

```{r echo=FALSE}
eigenvec <- pcs.cortex$rotation
head(eigenvec)
identical(rownames(eigenvec),rownames(cortexDes.uncor))
```

Perform ANOVA for categorical covariables and spearman correlations for continuous covariables (p-values have been corrected for multiple testing by FDR):

```{r echo=FALSE, warning=FALSE}
## create results table
eigen.results <- matrix(NA, nrow(eigenvec), 10)
dim(eigen.results) #nrow = 13   ncol = 10
rownames(eigen.results)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results) <- c("Chip", "Row", "Sex", "Disease_status", "HD_CAG_length","WT_CAG_length", "Age_Death", "Age_Onset", "Neuron", "PMI")
eigen.results <- as.data.frame(eigen.results)



### for-loop to do tests (eigenvalue ~ variable)

for(i in 1:ncol(eigenvec)){
  ### categorical data


  # Chip
  fit <- summary(aov(eigenvec[,i]~cortexDes.uncor$Sentrix_ID))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Chip[i] <- result
  
  # Row
  fit <- summary(aov(eigenvec[,i]~cortexDes.uncor$Row))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Row[i] <- result
  
  # Sex
  fit <- summary(aov(eigenvec[,i]~cortexDes.uncor$Sex))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Sex[i] <- result
  
  # Disease_status
  fit <- summary(aov(eigenvec[,i]~cortexDes.uncor$Disease_status))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Disease_status[i] <- result
  

  ### continuous data
  # HD_CAG_length
  eigen.results$HD_CAG_length[i] <- cor.test(eigenvec[,i],cortexDes.uncor$HD_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # WT_CAG_length
  eigen.results$WT_CAG_length[i] <- cor.test(eigenvec[,i],cortexDes.uncor$WT_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Age_Death
  eigen.results$Age_Death[i] <- cor.test(eigenvec[,i],cortexDes.uncor$Age_Death, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Age_Onset
  eigen.results$Age_Onset[i] <- cor.test(eigenvec[,i],cortexDes.uncor$Age_Onset, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Neuron
  eigen.results$Neuron[i] <- cor.test(eigenvec[,i],cortexDes.uncor$neuron, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # PMI
  eigen.results$PMI[i] <- cor.test(eigenvec[,i],cortexDes.uncor$PMI, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
}

## create corrected results table
eigen.results.cor <- matrix(NA, nrow(eigenvec), 10)
dim(eigen.results.cor) #nrow = 13   ncol = 10
rownames(eigen.results.cor)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results.cor) <- c("Chip", "Row", "Sex", "Disease_status", "HD_CAG_length","WT_CAG_length", "Age_Death", "Age_Onset", "Neuron","PMI")
eigen.results.cor <- as.data.frame(eigen.results.cor)

for(i in 1:ncol(eigen.results)){
  eigen.results.cor[,i]<-p.adjust(eigen.results[,i], method="BH")
  }
print(eigen.results.cor)
print(adj.prop.var)
```

We can see that with the normalized, ComBat-corrected, BEFORE CETS-correction cortex-only dataset (after multiple test correction, negating PC1 so all the PC numbers decrease by 1):
- Neuron/glial cell proportion is correlated with PC2 (ie actually PC1 accounting for 24.1% of the variance)
- Chip correlates with PC13 (ie actually PC12 accounting for 2.55%; this is negligible---refer to ComBat script for further info)

We will now visually assess if the neuronal/glial proportion (cell-type) effects have been removed from the cortex-only dataset:
```{r}
neuron<-cortexDes.uncor$neuron
disease<-cortexDes.uncor$Disease_status
pca.before.correction<-as.data.frame(pcs.cortex$rotation)
ggplot(pca.before.correction,aes(PC2,PC3,color=neuron, shape = disease))+geom_point(size=7)+theme_bw() + xlab("PC1 (24.1% of variance)") + ylab("PC2 (11.4% of variance") + scale_colour_gradient()
```

We can see that before CETS cell type correction, there is a strong evidence of brain cell type (neuron vs glia) effects in PC1 (accounting for 24.1% of the variance in the cortex-only normalized methylation data). 

#### For normalized, ComBat-corrected, AFTER CETS-correction cortex-only dataset:

```{r echo=FALSE}
pcs.cortex.cor<-prcomp(exprs(cortexDat.cor), center=TRUE, scale=FALSE) # scaling is not necessary for normalized datasets
```

```{r echo=FALSE}
plot(pcs.cortex.cor)
```

The figure shown is a scree plot which depicts the proportion of variance in the dataset that are explained by each of the principal components. We can see that the first principal component explains most of the variance in the methylation dataset. 

```{r echo=FALSE}
(pc.cortex.sum<-summary(pcs.cortex.cor))
```

We can see that PC1 explains 99.53% of the variance of the dataset.  However, we are only interested in the remaining PCs since the first PC just represents probe offset. We can re-adjust the proportion of variance explained by all subsequent PCs as follows: 


```{r echo=FALSE}
sum.prop<-pc.cortex.sum$importance
top.PC.prop<-sum.prop[2,1]
adj.sum.prop<-apply(sum.prop, 1, function(x) x/(1-top.PC.prop))
adj.prop.var<-adj.sum.prop[2:ncol(sum.prop),2]
adj.prop.var<-data.frame(adj.prop.var)
adj.prop.var$PC<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
rownames(adj.prop.var)<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
adj.prop.var$PC <- factor(adj.prop.var$PC, levels = paste("PC", 1:(ncol(sum.prop)-1), sep=""))
ggplot(adj.prop.var, aes(x=PC, y=adj.prop.var)) + geom_bar(stat="identity",fill="forestgreen") + theme_bw() + ylab("Proportion of Variance")
```

We will now correlate the known covariates with the PCs (note: we will also include the neuronal cell proportions calculated in the CETS brain cell type correction script). 

```{r echo=FALSE}
# convert categorical data into factors => do ANOVA
cortexDes.cor$Sentrix_ID <- as.factor(cortexDes.cor$Sentrix_ID)
cortexDes.cor$Sentrix_Position <- as.character(cortexDes.cor$Sentrix_Position)
cortexDes.cor$Sample_Group <- as.factor(cortexDes.cor$Sample_Group)
cortexDes.cor$Sex<- as.factor(cortexDes.cor$Sex)
cortexDes.cor$Disease_status <- as.factor(cortexDes.cor$Disease_status)
prop<-read.table("Cortex_CETS_Neuronal_Glia_Proportions.txt")
identical(rownames(cortexDes.cor),rownames(prop))
cortexDes.cor$neuron<-prop$neuron
cortexDes.cor$glia<-prop$glia
# for Row
for (i in 1:nrow(cortexDes.cor)){
  cortexDes.cor$Row[i]<-paste(substr(cortexDes.cor[i,"Sentrix_Position"], start=1, stop=3))
}
cortexDes.cor$Row<- as.factor(cortexDes.uncor$Row)
str(cortexDes.cor)
# continuous data => correlation (do spearman correlation because data is not always normally distributed)
```

Here are the eigenvectors (principal components) for the first few samples in the dataset:

```{r echo=FALSE}
eigenvec <- pcs.cortex.cor$rotation
head(eigenvec)
identical(rownames(eigenvec),rownames(cortexDes.cor))
```

Perform ANOVA for categorical covariables and spearman correlations for continuous covariables (p-values have been corrected for multiple testing by FDR):

```{r echo=FALSE, warning=FALSE}
## create results table
eigen.results <- matrix(NA, nrow(eigenvec), 10)
dim(eigen.results) #nrow = 13   ncol = 9
rownames(eigen.results)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results) <- c("Chip", "Row", "Sex", "Disease_status", "HD_CAG_length","WT_CAG_length", "Age_Death", "Age_Onset", "Neuron", "PMI")
eigen.results <- as.data.frame(eigen.results)


### for-loop to do tests (eigenvalue ~ variable)

for(i in 1:ncol(eigenvec)){
  ### categorical data


  # Chip
  fit <- summary(aov(eigenvec[,i]~cortexDes.cor$Sentrix_ID))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Chip[i] <- result
  
  # Row
  fit <- summary(aov(eigenvec[,i]~cortexDes.cor$Row))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Row[i] <- result
  
  # Sex
  fit <- summary(aov(eigenvec[,i]~cortexDes.cor$Sex))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Sex[i] <- result
  
  # Disease_status
  fit <- summary(aov(eigenvec[,i]~cortexDes.cor$Disease_status))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Disease_status[i] <- result
  

  ### continuous data
  # HD_CAG_length
  eigen.results$HD_CAG_length[i] <- cor.test(eigenvec[,i],cortexDes.cor$HD_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # WT_CAG_length
  eigen.results$WT_CAG_length[i] <- cor.test(eigenvec[,i],cortexDes.cor$WT_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Age_Death
  eigen.results$Age_Death[i] <- cor.test(eigenvec[,i],cortexDes.cor$Age_Death, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Age_Onset
  eigen.results$Age_Onset[i] <- cor.test(eigenvec[,i],cortexDes.cor$Age_Onset, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Neuron
  eigen.results$Neuron[i] <- cor.test(eigenvec[,i],cortexDes.cor$neuron, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # PMI
  eigen.results$PMI[i] <- cor.test(eigenvec[,i],cortexDes.cor$PMI, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
}

## create corrected results table
eigen.results.cor <- matrix(NA, nrow(eigenvec), 10)
dim(eigen.results.cor) #nrow = 13   ncol = 10
rownames(eigen.results.cor)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results.cor) <- c("Chip", "Row", "Sex", "Disease_status", "HD_CAG_length","WT_CAG_length", "Age_Death", "Age_Onset", "Neuron", "PMI")
eigen.results.cor <- as.data.frame(eigen.results.cor)

for(i in 1:ncol(eigen.results)){
  eigen.results.cor[,i]<-p.adjust(eigen.results[,i], method="BH")
  }
print(eigen.results.cor)
print(adj.prop.var)
```

We can see that with the normalized, ComBat-corrected, AFTER CETS-correction cortex-only dataset (after multiple test correction, negating PC1 so all the PC numbers decrease by 1):
- Neuron/glial cell proportion is correlated with PC13 (ie actually PC12 accounting for 0.64% of the variance). Note that this explains substantially less variance than PC2 which was correlated with neuron proportion in the normalized, Combat-corrected cortex-only dataset before CETS-correction. This indicates that the CETS-correction for neuronal/glial proportions was largely successful in the cortex-only dataset.
- Sex correlates with PC3 (ie actually PC2 accounting for 11.5% of the variance). This indicates that we need to adjust for sex in our linear models
- Chip correlates with PC12 (actually PC11 accounting for 3.4% of the variance). This is still negligible.

### PCA on HD cases of Cortex-Only Dataset

#### For HD cases in normalized, ComBat-corrected, AFTER CETS-correction cortex-only dataset:

```{r echo=FALSE}
# subset out HD cases from cortexDat.cor
HD.cases<-grep("HD", colnames(cortexDat.cor))
# input updated design matrix into cortexDat.cor methylumi object
updated.meta<-read.csv(file="cortexDes.final.csv", header=T, row.names=1)
identical(rownames(updated.meta), colnames(cortexDat.cor))
pData(cortexDat.cor)<-updated.meta
# subset out HD cases
HD.cortexDat.cor<-cortexDat.cor[,HD.cases]
colnames(HD.cortexDat.cor)
str(pData(HD.cortexDat.cor))
HD.cortexDes.cor<-pData(HD.cortexDat.cor)
# run PCA
pcs.hd.cortex<-prcomp(exprs(HD.cortexDat.cor), center=TRUE, scale=FALSE) # scaling is not necessary for normalized datasets
```

```{r echo=FALSE}
plot(pcs.hd.cortex)
```

The figure shown is a scree plot which depicts the proportion of variance in the dataset that are explained by each of the principal components. We can see that the first principal component explains most of the variance in the methylation dataset. 

```{r echo=FALSE}
(pc.hd.cortex.sum<-summary(pcs.hd.cortex))
```

We can see that PC1 explains 99.41% of the variance of the dataset.  However, we are only interested in the remaining PCs since the first PC just represents probe offset. We can re-adjust the proportion of variance explained by all subsequent PCs as follows: 


```{r echo=FALSE}
sum.prop<-pc.hd.cortex.sum$importance
top.PC.prop<-sum.prop[2,1]
adj.sum.prop<-apply(sum.prop, 1, function(x) x/(1-top.PC.prop))
adj.prop.var<-adj.sum.prop[2:ncol(sum.prop),2]
adj.prop.var<-data.frame(adj.prop.var)
adj.prop.var$PC<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
rownames(adj.prop.var)<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
adj.prop.var$PC <- factor(adj.prop.var$PC, levels = paste("PC", 1:(ncol(sum.prop)-1), sep=""))
ggplot(adj.prop.var, aes(x=PC, y=adj.prop.var)) + geom_bar(stat="identity",fill="purple") + theme_bw() + ylab("Proportion of Variance")
```

We will now correlate the known covariates with the PCs (note: we will also include the neuronal cell proportions calculated in the CETS brain cell type correction script). 

```{r echo=FALSE}
# convert categorical data into factors => do ANOVA
str(HD.cortexDes.cor)
HD.cortexDes.cor$Sentrix_ID <- as.factor(HD.cortexDes.cor$Sentrix_ID)
HD.cortexDes.cor$Sentrix_Position <- as.character(HD.cortexDes.cor$Sentrix_Position)
# for Row
for (i in 1:nrow(HD.cortexDes.cor)){
  HD.cortexDes.cor$Row[i]<-paste(substr(HD.cortexDes.cor[i,"Sentrix_Position"], start=1, stop=3))
}
HD.cortexDes.cor$Row<- as.factor(HD.cortexDes.cor$Row)
str(HD.cortexDes.cor)

# continuous data => correlation (do spearman correlation because data is not always normally distributed)
```

Here are the eigenvectors (principal components) for the first few samples in the dataset:

```{r echo=FALSE}
eigenvec <- pcs.hd.cortex$rotation
head(eigenvec)
identical(rownames(eigenvec),rownames(HD.cortexDes.cor))
```

Perform ANOVA for categorical covariables and spearman correlations for continuous covariables (p-values have been corrected for multiple testing by FDR):

```{r echo=FALSE, warning=FALSE}
## create results table
eigen.results <- matrix(NA, nrow(eigenvec), 10)
dim(eigen.results) #nrow = 7   ncol = 10
rownames(eigen.results)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results) <- c("Chip", "Row", "Sex", "Percentile_AO", "HD_CAG_length","WT_CAG_length", "Age_Death", "Age_Onset", "PMI", "HD_years")
eigen.results <- as.data.frame(eigen.results)
head(eigen.results)


### for-loop to do tests (eigenvalue ~ variable)

for(i in 1:ncol(eigenvec)){
  ### categorical data


  # Chip
  fit <- summary(aov(eigenvec[,i]~HD.cortexDes.cor$Sentrix_ID))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Chip[i] <- result
  
  # Row
  fit <- summary(aov(eigenvec[,i]~HD.cortexDes.cor$Row))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Row[i] <- result
  
  # Sex
  fit <- summary(aov(eigenvec[,i]~HD.cortexDes.cor$Sex))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Sex[i] <- result
  
  # Percentile_AO
  fit <- summary(aov(eigenvec[,i]~HD.cortexDes.cor$Percentile_AO))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Percentile_AO[i] <- result
  

  ### continuous data
  # HD_CAG_length
  eigen.results$HD_CAG_length[i] <- cor.test(eigenvec[,i],HD.cortexDes.cor$HD_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # WT_CAG_length
  eigen.results$WT_CAG_length[i] <- cor.test(eigenvec[,i],HD.cortexDes.cor$WT_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Age_Death
  eigen.results$Age_Death[i] <- cor.test(eigenvec[,i],HD.cortexDes.cor$Age_Death, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Age_Onset
  eigen.results$Age_Onset[i] <- cor.test(eigenvec[,i],HD.cortexDes.cor$Age_Onset, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # PMI
  eigen.results$PMI[i] <- cor.test(eigenvec[,i],HD.cortexDes.cor$PMI, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # HD_years
  eigen.results$HD_years[i] <- cor.test(eigenvec[,i],HD.cortexDes.cor$HD_years, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
}

## create corrected results table
eigen.results.cor <- matrix(NA, nrow(eigenvec), 10)
dim(eigen.results.cor) #nrow = 7   ncol = 10
rownames(eigen.results.cor)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results.cor) <- c("Chip", "Row", "Sex", "Percentile_AO", "HD_CAG_length","WT_CAG_length", "Age_Death", "Age_Onset", "PMI", "HD_years")
eigen.results.cor <- as.data.frame(eigen.results.cor)

for(i in 1:ncol(eigen.results)){
  eigen.results.cor[,i]<-p.adjust(eigen.results[,i], method="BH")
  }
print(eigen.results.cor)
print(adj.prop.var)
```

We can see that with the HD cases of the normalized, ComBat-corrected, AFTER CETS-correction cortex-only dataset (after multiple test correction, negating PC1 so all the PC numbers decrease by 1):
- none of the tested metavariables were significantly associated with any of the PCs

 
### For matched dataset

#### For normalized, ComBat-corrected, BEFORE CETS-correction matched dataset:
```{r echo=FALSE}
pcs.matched<-prcomp(exprs(matchedDat.uncor), center=TRUE, scale=FALSE) # scaling is not necessary for a normalized dataset
```

```{r echo=FALSE}
plot(pcs.matched)
```

The figure shown is a scree plot which depicts the proportion of variance in the dataset that are explained by each of the principal components. We can see that the first principal component explains most of the variance in the methylation dataset. 

```{r echo=FALSE}
(pc.sum.matched<-summary(pcs.matched))
```

We can see that PC1 explains 96.57% of the variance of the dataset. However, we are only interested in the remaining PCs since the first PC just represents probe offset. We can re-adjust the proportion of variance explained by all subsequent PCs as follows: 

```{r echo=FALSE}
sum.prop<-pc.sum.matched$importance
top.PC.prop<-sum.prop[2,1]
adj.sum.prop<-apply(sum.prop, 1, function(x) x/(1-top.PC.prop))
adj.prop.var<-adj.sum.prop[2:ncol(sum.prop),2]
adj.prop.var<-data.frame(adj.prop.var)
adj.prop.var$PC<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
rownames(adj.prop.var)<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
adj.prop.var$PC <- factor(adj.prop.var$PC, levels = paste("PC", 1:(ncol(sum.prop)-1), sep=""))
ggplot(adj.prop.var, aes(x=PC, y=adj.prop.var)) + geom_bar(stat="identity",fill="goldenrod3") + theme_bw() + ylab("Proportion of Variance")
```

```{r echo=FALSE}

# convert categorical data into factors => do ANOVA
matchedDes.uncor$Sentrix_ID <- as.factor(matchedDes.uncor$Sentrix_ID)
matchedDes.uncor$Sentrix_Position <- as.character(matchedDes.uncor$Sentrix_Position)
matchedDes.uncor$Sample_Group <- as.factor(matchedDes.uncor$Sample_Group)
matchedDes.uncor$Sex<- as.factor(matchedDes.uncor$Sex)
matchedDes.uncor$Disease_status <- as.factor(matchedDes.uncor$Disease_status)
for (i in 1:nrow(matchedDes.uncor)){
  matchedDes.uncor$Row[i]<-paste(substr(matchedDes.uncor[i,"Sentrix_Position"], start=1, stop=3))
}
matchedDes.uncor$Row<- as.factor(matchedDes.uncor$Row)
str(matchedDes.uncor)
# continuous data => correlation (do spearman correlation because data is not always normally distributed)
```

Here are the eigenvectors (principal components) for the first few samples in the dataset:

```{r echo=FALSE}
eigenvec <- pcs.matched$rotation
head(eigenvec)
identical(rownames(eigenvec),rownames(matchedDes.uncor))
```

Perform ANOVA for categorical covariables and spearman correlations for continuous covariables (p-values have been corrected for multiple testing by FDR):

```{r echo=FALSE, warning=FALSE}
## create results table (exclude chip since all of these samples were run on a single chip)
eigen.results <- matrix(NA, nrow(eigenvec), 8)
dim(eigen.results) #nrow = 10   ncol = 8
rownames(eigen.results)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results) <- c("Tissue", "Row", "Sex", "Disease_status", "HD_CAG_length","WT_CAG_length", "Age_Death", "PMI")
colnames(eigen.results)
eigen.results <- as.data.frame(eigen.results)
head(eigen.results)


### for-loop to do tests (eigenvalue ~ variable)

for(i in 1:ncol(eigenvec)){
  ### categorical data
 # Tissue
  fit <- summary(aov(eigenvec[,i]~matchedDes.uncor$Sample_Group))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Tissue[i] <- result
  
  # Row
  fit <- summary(aov(eigenvec[,i]~matchedDes.uncor$Row))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Row[i] <- result
  
  # Sex
  fit <- summary(aov(eigenvec[,i]~matchedDes.uncor$Sex))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Sex[i] <- result
  
  # Disease_status
  fit <- summary(aov(eigenvec[,i]~matchedDes.uncor$Disease_status))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Disease_status[i] <- result
  

  ### continuous data
  # HD_CAG_length
  eigen.results$HD_CAG_length[i] <- cor.test(eigenvec[,i],matchedDes.uncor$HD_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # WT_CAG_length
  eigen.results$WT_CAG_length[i] <- cor.test(eigenvec[,i],matchedDes.uncor$WT_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Age_Death
  eigen.results$Age_Death[i] <- cor.test(eigenvec[,i],matchedDes.uncor$Age_Death, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # PMI
  eigen.results$PMI[i] <- cor.test(eigenvec[,i],matchedDes.uncor$PMI, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
}


## Perform multiple test correction using Bioconductor qvalue package (FDR)
## create corrected results table
eigen.results.cor <- matrix(NA, nrow(eigenvec), 8)
dim(eigen.results.cor) #nrow = 10   ncol = 8
rownames(eigen.results.cor)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results.cor) <- c( "Tissue","Row", "Sex", "Disease_status", "HD_CAG_length","WT_CAG_length", "Age_Death", "PMI")
colnames(eigen.results.cor)
eigen.results.cor <- as.data.frame(eigen.results.cor)

for(i in 1:ncol(eigen.results)){
  qval<-qvalue(eigen.results[,i])
  eigen.results.cor[,i]<-qval$qvalues
}
print(eigen.results.cor)
print(adj.prop.var)
```

We can see that with the normalized, ComBat-corrected, BEFORE CETS-correction matched dataset (after multiple test correction, negating PC1 so all the PC numbers decrease by 1):
- Tissue (liver vs cortex) correlates with PC2 (ie actually PC1 accounting for 80.3% of the variance)
- None of the other covariates are significantly correlated with any of the other PCs
 
#### For normalized, ComBat-corrected, AFTER CETS-correction matched dataset:
```{r echo=FALSE}
pcs.matched.cor<-prcomp(exprs(matchedDat.cor), center=TRUE, scale=FALSE) # scaling is not necessary for a normalized dataset
```

```{r echo=FALSE}
plot(pcs.matched.cor)
```

The figure shown is a scree plot which depicts the proportion of variance in the dataset that are explained by each of the principal components. We can see that the first principal component explains most of the variance in the methylation dataset. 

```{r echo=FALSE}
(pc.sum.matched<-summary(pcs.matched.cor))
```

We can see that PC1 explains 96.57% of the variance of the dataset. However, we are only interested in the remaining PCs since the first PC just represents probe offset. We can re-adjust the proportion of variance explained by all subsequent PCs as follows: 

```{r echo=FALSE}
sum.prop<-pc.sum.matched$importance
top.PC.prop<-sum.prop[2,1]
adj.sum.prop<-apply(sum.prop, 1, function(x) x/(1-top.PC.prop))
adj.prop.var<-adj.sum.prop[2:ncol(sum.prop),2]
adj.prop.var<-data.frame(adj.prop.var)
adj.prop.var$PC<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
rownames(adj.prop.var)<-paste("PC", 1:(ncol(sum.prop)-1), sep="")
adj.prop.var$PC <- factor(adj.prop.var$PC, levels = paste("PC", 1:(ncol(sum.prop)-1), sep=""))
ggplot(adj.prop.var, aes(x=PC, y=adj.prop.var)) + geom_bar(stat="identity",fill="orange") + theme_bw() + ylab("Proportion of Variance")
```

```{r echo=FALSE}

# convert categorical data into factors => do ANOVA
matchedDes.cor$Sentrix_ID <- as.factor(matchedDes.cor$Sentrix_ID)
matchedDes.cor$Sentrix_Position <- as.character(matchedDes.cor$Sentrix_Position)
matchedDes.cor$Sample_Group <- as.factor(matchedDes.cor$Sample_Group)
matchedDes.cor$Sex<- as.factor(matchedDes.cor$Sex)
matchedDes.cor$Disease_status <- as.factor(matchedDes.cor$Disease_status)
for (i in 1:nrow(matchedDes.cor)){
  matchedDes.cor$Row[i]<-paste(substr(matchedDes.cor[i,"Sentrix_Position"], start=1, stop=3))
}
matchedDes.cor$Row<- as.factor(matchedDes.cor$Row)
str(matchedDes.cor)
# continuous data => correlation (do spearman correlation because data is not always normally distributed)
```

Here are the eigenvectors (principal components) for the first few samples in the dataset:

```{r echo=FALSE}
eigenvec <- pcs.matched.cor$rotation
head(eigenvec)
identical(rownames(eigenvec),rownames(matchedDes.cor))
```

Perform ANOVA for categorical covariables and spearman correlations for continuous covariables (p-values have been corrected for multiple testing by FDR):

```{r echo=FALSE, warning=FALSE}
## create results table (exclude chip since all of these samples were run on a single chip)
eigen.results <- matrix(NA, nrow(eigenvec), 8)
dim(eigen.results) #nrow = 10   ncol = 8
rownames(eigen.results)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results) <- c("Tissue","Row", "Sex", "Disease_status", "HD_CAG_length","WT_CAG_length", "Age_Death", "PMI")
colnames(eigen.results)
eigen.results <- as.data.frame(eigen.results)


### for-loop to do tests (eigenvalue ~ variable)

for(i in 1:ncol(eigenvec)){
  ### categorical data
 # Tissue
  fit <- summary(aov(eigenvec[,i]~matchedDes.cor$Sample_Group))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Tissue[i] <- result
  
  # Row
  fit <- summary(aov(eigenvec[,i]~matchedDes.cor$Row))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Row[i] <- result
  ## keep getting following error with this code: Error in eigen.results$Chip_Position[i] <- result : replacement has length zero
  
  # Sex
  fit <- summary(aov(eigenvec[,i]~matchedDes.cor$Sex))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Sex[i] <- result
  
  # Disease_status
  fit <- summary(aov(eigenvec[,i]~matchedDes.cor$Disease_status))
  result <- fit[[1]]$"Pr(>F)"[1] 
  eigen.results$Disease_status[i] <- result
  

  ### continuous data
  # HD_CAG_length
  eigen.results$HD_CAG_length[i] <- cor.test(eigenvec[,i],matchedDes.cor$HD_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # WT_CAG_length
  eigen.results$WT_CAG_length[i] <- cor.test(eigenvec[,i],matchedDes.cor$WT_CAG_length, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # Age_Death
  eigen.results$Age_Death[i] <- cor.test(eigenvec[,i],matchedDes.cor$Age_Death, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
  
  # PMI
  eigen.results$PMI[i] <- cor.test(eigenvec[,i],matchedDes.cor$PMI, alternative = "two.sided", method="spearman", na.action=na.omit)$p.value
}


## Perform multiple test correction using Bioconductor qvalue package (FDR)
## create corrected results table
eigen.results.cor <- matrix(NA, nrow(eigenvec), 8)
dim(eigen.results.cor) #nrow = 10   ncol = 8
rownames(eigen.results.cor)<-paste("PC", 1:nrow(eigenvec), sep="")
colnames(eigen.results.cor) <- c( "Tissue", "Row", "Sex", "Disease_status", "HD_CAG_length","WT_CAG_length", "Age_Death", "PMI")
colnames(eigen.results.cor)
eigen.results.cor <- as.data.frame(eigen.results.cor)

for(i in 1:ncol(eigen.results)){
  qval<-qvalue(eigen.results[,i])
  eigen.results.cor[,i]<-qval$qvalues
}
print(eigen.results.cor)
print(adj.prop.var)
```

We can see that with the normalized, ComBat-corrected, AFTER CETS-correction matched dataset (after multiple test correction, negating PC1 so all the PC numbers decrease by 1):
- Tissue (liver vs cortex) correlates with PC2 (ie actually PC1 accounting for 80.45% of the variance)
- Sex correlates with PC4 (ie actually PC3 accounting for 3.3% of the variance)
- Age of death correlates with PC10 (ie actually PC9 accounting for 0.17% of the variance). 

We will now visually assess the tissue effects in PC1 in the corrected matched dataset:
```{r}
Tissue<-factor(matchedDes.cor$Sample_Group)
pca.after.correction<-as.data.frame(pcs.matched.cor$rotation)
ggplot(pca.after.correction,aes(PC2,PC3, color=Tissue))+geom_point(shape=19, size=7)+theme_bw() + xlab("PC1 (80.5% of variance)") + ylab("PC2 (4.2% of variance)") + scale_color_manual(values=c("darkgoldenrod1", "blue"))
```

We will now visually assess the sex effects in PC3 in the corrected matched dataset:
```{r}
Sex<-factor(matchedDes.cor$Sex)
pca.after.correction<-as.data.frame(pcs.matched.cor$rotation)
ggplot(pca.after.correction,aes(PC4,PC5, color=Sex))+geom_point(shape=19, size=7)+theme_bw() + xlab("PC3 (3.3% of variance)") + ylab("PC4 (3.0% of variance)")
```

We will now visually assess the age of onset and age of death in PC9 in the corrected matched dataset:
```{r}
Age.Death<-matchedDes.cor$Age_Death
pca.after.correction<-as.data.frame(pcs.matched.cor$rotation)
ggplot(pca.after.correction,aes(PC9,PC10, size=Age.Death))+geom_point(shape=19, color="firebrick")+theme_bw() + xlab("PC8 (1.9% of variance)") + ylab("PC9 (0.17% of variance)")

Age.Onset<-matchedDes.cor$Age_Onset
pca.after.correction<-as.data.frame(pcs.matched.cor$rotation)
ggplot(pca.after.correction,aes(PC9,PC10, size=Age.Onset))+geom_point(shape=17,color="darkmagenta")+theme_bw() + xlab("PC8 (1.9% of variance)") + ylab("PC9 (0.17% of variance)")
```
