==> 8_HTT_Leavitt_Matched_Pyrosequencing_Validation.Rmd <==
HTT_Leavitt Pyrosequencing vs 450K Verification
========================================================
## By: Sumaiya Islam
## Date: April 30, 2015

### Script contents:
- Correlation of pyrosequencing and 450K differential CpG sites (cg07240470,cg11324953,cg15544235). Note that these pyrosequencing assays were performed by Lisa (Kobor lab) in November 2014. 

### A. Set up wd & packages

We will initially set our working directory and load our libraries.
```{r HTT_Leavitt_setup, include = FALSE, echo=FALSE}
setwd("/home/sislam/HTT_Leavitt")
library(methylumi)
library(gplots)
library(marray)
library(lumi)
library(lattice)
library("RColorBrewer")
library(knitr)
library(xtable)
library(limma)
library(reshape)
library(ggplot2)
library(grid)
library(gridExtra)
library(MASS)
```

### B. Load data

We will load the data containing both the beta values for both the 450K and pyrosequencing platforms for the target CpGs (cg07240470,cg11324953,cg15544235). 

```{r load_data}
validation<-read.csv(file="Validation_Dataset_Betas.csv", header = T, row.names=1, check.names=F)
head(validation)
str(validation)
colnames(validation)

# Calculate diff (array-pyro measures) and average (array + pyro/2) for each sample in the validation dataset (for Bland-Altman plot)

for (i in 1:nrow(validation)){
  validation$diff.cg07240470<-(validation[,"array_cg07240470"]-validation[,"pyro_cg07240470"])
  validation$avg.cg07240470<-((validation[,"array_cg07240470"]+validation[,"pyro_cg07240470"])/2)
  validation$diff.cg11324953<-(validation[,"array_cg11324953"]-validation[,"pyro_cg11324953"])
  validation$avg.cg11324953<-((validation[,"array_cg11324953"]+validation[,"pyro_cg11324953"])/2)
  validation$diff.cg15544235<-(validation[,"array_cg15544235"]-validation[,"pyro_cg15544235"])
  validation$avg.cg15544235<-((validation[,"array_cg15544235"]+validation[,"pyro_cg15544235"])/2)  
  }
```

### C. Correlations between Array and Pyrosequencing Measures

We will correlate the data containing both the beta values for both the 450K and pyrosequencing platforms for the target CpGs (cg07240470,cg11324953,cg15544235). The resulting correlations will be plotted in side-by-side scatterplots and Bland-Altman plots. 

```{r}

# for cg07240470

## Spearman's Correlation
cor(validation$array_cg07240470, validation$pyro_cg07240470, method="spearman") # spearman's rho = 0.6190476
cor.test(validation$array_cg07240470, validation$pyro_cg07240470, alternative="two.sided", method="spearman") # spearman's rho = 0.6190476 p = 0.115
## Scatterplot
scatter.cg07240470<-ggplot(validation, aes(array_cg07240470, pyro_cg07240470)) + geom_point(aes(colour = Tissue),size=6) + geom_smooth(method=lm, se=FALSE, fill="darkgrey", colour="black", size=2) + scale_color_manual(values=c("orange", "blue")) + ylab("cg07240470 Pyroseqencing") + xlab("cg07240470 450K Array") + theme_bw() + annotate("text", x = 0.765, y = 1, label = "Spearman's rho = 0.619") 
## Bland-Altman plot
m<-mean(validation$diff.cg07240470)
upper<-m+2*(sd(validation$diff.cg07240470))
lower<-m-2*(sd(validation$diff.cg07240470))
BAplot.cg07240470<-ggplot(validation, aes(avg.cg07240470, diff.cg07240470)) + geom_point(aes(colour = Tissue), size=6) + geom_hline(yintercept=c(m, upper, lower)) + scale_color_manual(values=c("orange", "blue")) + theme_bw() + ylab("Diff (Array - Pyro)") + xlab("Average of Array and Pyro Measures") # lines represent mean difference + or - 2sd

grid.arrange(scatter.cg07240470,BAplot.cg07240470, ncol=2)

# for cg11324953

## Spearman's Correlation
cor(validation$array_cg11324953, validation$pyro_cg11324953, method="spearman") # spearman's rho = 0.7142857
cor.test(validation$array_cg11324953, validation$pyro_cg11324953, , alternative="two.sided", method="spearman") # spearman's rho = 0.7142857 p = 0.058
## Scatterplot
scatter.cg11324953<-ggplot(validation, aes(array_cg11324953, pyro_cg11324953)) + geom_point(aes(colour = Tissue), size=6) + geom_smooth(method=lm, se=FALSE, fill="darkgrey", colour="black", size=2) + scale_color_manual(values=c("orange", "blue")) + ylab("cg11324953 Pyroseqencing") + xlab("cg11324953 450K Array") + theme_bw() + annotate("text", x = 0.6, y = 0.8, label = "Spearman's rho = 0.714") 
## Bland-Altman plot
m<-mean(validation$diff.cg11324953)
upper<-m+2*(sd(validation$diff.cg11324953))
lower<-m-2*(sd(validation$diff.cg11324953))
BAplot.cg11324953<-ggplot(validation, aes(avg.cg11324953, diff.cg11324953)) + geom_point(aes(colour = Tissue), size=6) + geom_hline(yintercept=c(m, upper, lower)) + scale_color_manual(values=c("orange", "blue")) + theme_bw() + ylab("Diff (Array - Pyro)") + xlab("Average of Array and Pyro Measures")
grid.arrange(scatter.cg11324953,BAplot.cg11324953, ncol=2)

# for cg15544235

## Spearman's Correlation
cor(validation$array_cg15544235, validation$pyro_cg15544235, method="spearman") # spearman's rho = 0.7857143 
cor.test(validation$array_cg15544235, validation$pyro_cg15544235, alternative = "two.sided", method="spearman") # spearman's rho = 0.7857143 p = 0.028
## Scatterplot
scatter.cg15544235<-ggplot(validation, aes(array_cg15544235, pyro_cg15544235)) + geom_point(aes(colour = Tissue), size=6) + geom_smooth(method=lm, se=FALSE, fill="darkgrey", colour="black", size=2) + scale_color_manual(values=c("orange", "blue")) + ylab("cg15544235 Pyroseqencing") + xlab("cg15544235 450K Array") + theme_bw() + annotate("text", x = 0.87, y = 0.875, label = "Spearman's rho = 0.786")  
## Bland-Altman plot
m<-mean(validation$diff.cg15544235)
upper<-m+2*(sd(validation$diff.cg15544235))
lower<-m-2*(sd(validation$diff.cg15544235))
BAplot.cg15544235<-ggplot(validation, aes(avg.cg15544235, diff.cg15544235)) + geom_point(aes(colour = Tissue), size=6) + geom_hline(yintercept=c(m, upper, lower)) + scale_color_manual(values=c("orange", "blue")) + theme_bw() + ylab("Diff (Array - Pyro)") + xlab("Average of Array and Pyro Measures")
grid.arrange(scatter.cg15544235,BAplot.cg15544235, ncol=2)
print(BAplot.cg15544235)
```

Check if the underlying cgs in each assay are correlated. 

```{r}
# load summarized pyro data
head(pv.dat<-read.csv("Pyro_Validation_Data.csv", header = T, row.names=1, check.names=F))
str(pv.dat)
pv.dec.dat<-data.frame(pv.dat[,1:2], apply(pv.dat[,3:12],2, function(x) x/100)) # convert methylation percentages to decimal values
# remove HDB129 as this individual in the validation set had hepatitis which likely affected the liver tissue
head(pv.dec.dat<-pv.dec.dat[c(-2,-10,-18),])
op.dec.dat<-subset(pv.dec.dat, Dataset=="Original")

# Scatterplot matrix of Spearman correlations for 

panel.lm<-function (x, y, col = par("col"), bg = NA, pch = par("pch"),
     cex = 1, col.lm = "red", lwd=par("lwd"), ...)
 {
     points(x, y, pch = 16, col="black", bg = bg, cex = cex)
     ok <- is.finite(x) & is.finite(y)
     if (any(ok))
         abline(lm(y[ok]~x[ok]), col = col.lm, ...)}

 panel.sse<- function(y, x, digits=2,...)
      {
         usr <- par("usr"); on.exit(par(usr))
         par(usr = c(0, 1, 0, 1))

         model <- summary(lm(y~x))
                              # r2<- model$r.squared
                              # r<-sqrt(r2)*sign(model$coef[2,1])
                              # p <- model$coef[2,4]
                              r <- cor(x,y,method="spearman", use="pairwise.complete.obs")
                              # print(r)
                              p <- cor.test(x,y,method="spearman",alternative="two.sided",conf.level=0.95)
                              p <- p$p.value
                              # print(p)

                                
         txt <- round(r, digits)
         txt <- bquote(r == .(txt))
  	cex.cor<-0.8/strwidth(txt)
         text(0.5, 0.7, txt, cex=1.2)

         # txt <- round(r2, digits)
         # txt <- bquote(r^2 == .(txt))
         # text(0.5, 0.5, txt, cex=1.5)

         txt <- round(p, digits)
         txt <- bquote(p == .(txt))
         text(0.5, 0.3, txt, cex=1.0)
}
     

## Plot pairwise correlations of methylation sites for each site (note: Illumina cg id provided for array site)
assay1.panel<-pairs(op.dec.dat[,c(3,4,5,6,7,8)],main="Assay 1",lower.panel=panel.sse,upper.panel=panel.lm) # all Assay 1 sites (only first 3 sites are correlated; also site 5 and 6 but neither are the target array site)
assay1.panel<-pairs(op.dec.dat[,c(3,4,5)],main="Assay 1",lower.panel=panel.sse,upper.panel=panel.lm) # only correlated sites of Assay 1
assay2.panel<-pairs(op.dec.dat[,c(9,10)],main="Assay 2",lower.panel=panel.sse,upper.panel=panel.lm) # all Assay 2 sites correlated
assay3.panel<-pairs(op.dec.dat[,c(11,12)],main="Assay 3",lower.panel=panel.sse,upper.panel=panel.lm)


## generate averaged methylation values for correlated sites in each pyro assay
identical(rownames(validation), rownames(op.dec.dat)) # False, need to re-order op.dec.dat to match order of samples in validation data frame
sample.order<-rownames(validation)
op.dec.dat.ordered<-op.dec.dat[match(sample.order, rownames(op.dec.dat)),] # re-order op.dec.dat to match order of samples in validation data frame
identical(rownames(validation), rownames(op.dec.dat.ordered)) # True
assay1.avg.values<-apply(op.dec.dat.ordered[,3:5], 1, mean)
assay2.avg.values<-apply(op.dec.dat.ordered[,9:10],1, mean)
assay3.avg.values<-apply(op.dec.dat.ordered[,11:12],1, mean)
validation$assay1.avg<-assay1.avg.values
validation$assay2.avg<-assay2.avg.values
validation$assay3.avg<-assay3.avg.values

### Explored how taking averaged methylation values for each pyro assay would impact correlation to array data but note that this is not verification of the array data so not needed for the manuscript
# # Calculate diff (array-pyro measures) and average (array + avg. pyro/2) for each sample in the validation dataset (for Bland-Altman plot)
# for (i in 1:nrow(validation)){
#   validation$diff.cg07240470.cor<-(validation[,"array_cg07240470"]-validation[,"assay1.avg"])
#   validation$avg.cg07240470.cor<-((validation[,"array_cg07240470"]+validation[,"assay1.avg"])/2)
#   validation$diff.cg11324953.cor<-(validation[,"array_cg11324953"]-validation[,"assay2.avg"])
#   validation$avg.cg11324953.cor<-((validation[,"array_cg11324953"]+validation[,"assay2.avg"])/2)
#   validation$diff.cg15544235.cor<-(validation[,"array_cg15544235"]-validation[,"assay3.avg"])
#   validation$avg.cg15544235.cor<-((validation[,"array_cg15544235"]+validation[,"assay3.avg"])/2)  
#   }
# 
# # For Assay 1
# 
# ## Spearman's Correlation
# cor(validation$array_cg07240470, validation$assay1.avg, method="spearman") # spearman's rho = 0.6190476
# cor.test(validation$array_cg07240470, validation$assay1.avg, alternative="two.sided", method="spearman") # spearman's rho = 0.6190476 p = 0.115
# ## Scatterplot
# scatter.cg07240470<-ggplot(validation, aes(array_cg07240470, assay1.avg)) + geom_point(aes(colour = Tissue),size=6) + geom_smooth(method=lm, se=FALSE, fill="darkgrey", colour="black", size=2) + scale_color_manual(values=c("orange", "blue")) + ylab("Pyroseqencing Average") + xlab("cg07240470 450K Array") + theme_bw() + annotate("text", x = 0.765, y = 1, label = "Spearman's rho = 0.619") 
# 
# 
# ## Bland-Altman plot
# m<-mean(validation$diff.cg07240470.cor)
# upper<-m+2*(sd(validation$diff.cg07240470.cor))
# lower<-m-2*(sd(validation$diff.cg07240470.cor))
# BAplot.cg07240470<-ggplot(validation, aes(avg.cg07240470.cor, diff.cg07240470.cor)) + geom_point(aes(colour = Tissue), size=6) + geom_hline(yintercept=c(m, upper, lower)) + scale_color_manual(values=c("orange", "blue")) + theme_bw() + ylab("Diff (Array - Pyro)") + xlab("Average of Array and Averaged Pyro Measures") # lines represent mean difference + or - 2sd
# 
# grid.arrange(scatter.cg07240470,BAplot.cg07240470, ncol=2)
# 
# # For Assay 2
# 
# ## Spearman's Correlation
# cor(validation$array_cg11324953, validation$assay2.avg, method="spearman") # spearman's rho = 0.7380952
# cor.test(validation$array_cg11324953, validation$assay2.avg, , alternative="two.sided", method="spearman") # spearman's rho = 0.7380952 p = 0.04583
# ## Scatterplot
# scatter.cg11324953<-ggplot(validation, aes(array_cg11324953, assay2.avg)) + geom_point(aes(colour = Tissue), size=6) + geom_smooth(method=lm, se=FALSE, fill="darkgrey", colour="black", size=2) + scale_color_manual(values=c("orange", "blue")) + ylab("Pyroseqencing Average") + xlab("cg11324953 450K Array") + theme_bw() + annotate("text", x = 0.6, y = 0.8, label = "Spearman's rho = 0.738") 
# ## Bland-Altman plot
# m<-mean(validation$diff.cg11324953.cor)
# upper<-m+2*(sd(validation$diff.cg11324953.cor))
# lower<-m-2*(sd(validation$diff.cg11324953.cor))
# BAplot.cg11324953<-ggplot(validation, aes(avg.cg11324953.cor, diff.cg11324953.cor)) + geom_point(aes(colour = Tissue), size=6) + geom_hline(yintercept=c(m, upper, lower)) + scale_color_manual(values=c("orange", "blue")) + theme_bw() + ylab("Diff (Array - Pyro)") + xlab("Average of Array and Averaged Pyro Measures")
# grid.arrange(scatter.cg11324953,BAplot.cg11324953, ncol=2)
# 
# # For Assay 3
# 
# ## Spearman's Correlation
# cor(validation$array_cg15544235, validation$assay3.avg, method="spearman") # spearman's rho = 0.8095238 
# cor.test(validation$array_cg15544235, validation$assay3.avg, alternative = "two.sided", method="spearman") # spearman's rho = 0.8095238 p = 0.02178
# ## Scatterplot
# scatter.cg15544235<-ggplot(validation, aes(array_cg15544235, assay3.avg)) + geom_point(aes(colour = Tissue), size=6) + geom_smooth(method=lm, se=FALSE, fill="darkgrey", colour="black", size=2) + scale_color_manual(values=c("orange", "blue")) + ylab("Pyroseqencing Average") + xlab("cg15544235 450K Array") + theme_bw() + annotate("text", x = 0.87, y = 0.875, label = "Spearman's rho = 0.810")  
# ## Bland-Altman plot
# m<-mean(validation$diff.cg15544235.cor)
# upper<-m+2*(sd(validation$diff.cg15544235.cor))
# lower<-m-2*(sd(validation$diff.cg15544235.cor))
# BAplot.cg15544235<-ggplot(validation, aes(avg.cg15544235.cor, diff.cg15544235.cor)) + geom_point(aes(colour = Tissue), size=6) + geom_hline(yintercept=c(m, upper, lower)) + scale_color_manual(values=c("orange", "blue")) + theme_bw() + ylab("Diff (Array - Pyro)") + xlab("Average of Array and Averaged Pyro Measures")
# grid.arrange(scatter.cg15544235,BAplot.cg15544235, ncol=2)
```

Test if averaged methylation values for each assay are significantly different between cortex and liver in both the original samples and validation samples, respectively
```{r}
## calculate avg methylation values for each assay across the samples
pv.dec.dat$assay1.avg.values<-apply(pv.dec.dat[,3:5], 1, mean)
pv.dec.dat$assay2.avg.values<-apply(pv.dec.dat[,9:10],1, mean)
pv.dec.dat$assay3.avg.values<-apply(pv.dec.dat[,11:12],1, mean)

## test differences between cortex and liver tissues in both original and validation samples using Wilcoxon paired signed-rank test
head(cor.liv.dat<-subset(pv.dec.dat, Tissue=="Cortex"|Tissue=="Liver"))
(assay.1.stat<-wilcox.test(cor.liv.dat[1:7,]$assay1.avg, cor.liv.dat[8:14,]$assay1.avg, paired=TRUE)) # for assay 1
(assay.2.stat<-wilcox.test(cor.liv.dat[1:7,]$assay2.avg, cor.liv.dat[8:14,]$assay2.avg, paired=TRUE)) # for assay 2
(assay.3.stat<-wilcox.test(cor.liv.dat[1:7,]$assay3.avg, cor.liv.dat[8:14,]$assay3.avg, paired=TRUE)) # for assay 3
(adjusted.p.values<-p.adjust(c(assay.1.stat$p.value,assay.2.stat$p.value,assay.3.stat$p.value), method="BH"))

## generate boxplots for each assay
# Assay 1
ggplot(cor.liv.dat, aes(Tissue, assay1.avg.values))+
  geom_boxplot(color="grey", outlier.shape=NA)+
  geom_point(aes(color=Dataset), shape=19, size=4, position = position_jitter(w = 0.2))+
  theme_bw()+xlab("Tissue")+ylab("Pyro Assay 1 Avg Methylation")+
  scale_color_manual(values=c("#99d594", "#fc8d59"))
# Assay 2
ggplot(cor.liv.dat, aes(Tissue, assay2.avg.values))+
  geom_boxplot(color="grey", outlier.shape=NA)+
  geom_point(aes(color=Dataset), shape=19, size=4, position = position_jitter(w = 0.2))+
  theme_bw()+xlab("Tissue")+ylab("Pyro Assay 2 Avg Methylation")+
  scale_color_manual(values=c("#99d594", "#fc8d59"))
# Assay 3
ggplot(cor.liv.dat, aes(Tissue, assay3.avg.values))+
  geom_boxplot(color="grey", outlier.shape=NA)+
  geom_point(aes(color=Dataset), shape=19, size=4, position = position_jitter(w = 0.2))+
  theme_bw()+xlab("Tissue")+ylab("Pyro Assay 3 Avg Methylation")+
  scale_color_manual(values=c("#99d594", "#fc8d59"))
```
